LayoutTests/platform/mac-lion/media/video-size-intrinsic-scale-expected.txt

Test size before movie is open:
EXPECTED (video.videoWidth == '0') OK
EXPECTED (video.videoHeight == '0') OK

Loading "content/scaled-matrix.mov". Movie saved with a non identity matrix.
EVENT(loadstart)
EVENT(durationchange)
EVENT(loadedmetadata)
EXPECTED (video.videoWidth == '640') OK
EXPECTED (video.videoHeight == '480') OK

Loading "content/720x576i-anamorphic.mov". Movie encoded with PAL non-square pixels, clean aperture corrects pixel aspect ratio when displayed.
EVENT(loadstart)
EVENT(durationchange)
EVENT(loadedmetadata)
EXPECTED (video.videoWidth == '1024'), OBSERVED '720' FAIL
EXPECTED (video.videoHeight == '576') OK

END OF TEST
